Move retry related logic to its own file
